refactor: create actions and store and reducers folder
