+++ src/actions/index.ts
@@ -1,0 +1,2 @@
+export { changeTheme } from "./theme.action";
+export { createUser } from "./user.action";
