+++ src/actions/theme.action.ts
@@ -1,0 +1,3 @@
+import { createAction } from "@reduxjs/toolkit";
+
+export const changeTheme = createAction<string>("theme/change");
