+++ src/actions/user.action.ts
@@ -1,0 +1,4 @@
+import type { User } from "@interfaces";
+import { createAction } from "@reduxjs/toolkit";
+
+export const createUser = createAction<User>("user/createUser");
