+++ src/reducers/index..ts
@@ -1,0 +1,2 @@
+export { default as themeReducer } from "./theme.reducer";
+export { default as userReducer } from "./user.reducer";
