+++ src/store/index.ts
@@ -1,0 +1,2 @@
+export type { Dispatch, RootState } from "./store";
+export { store } from "./store";
